+++ test/engine_showdown.dart
@@ -1,0 +1,22 @@
+import 'package:bishop/bishop.dart';
+
+main(List<String> args) async {
+  Game game = Game(variant: Variant.mini());
+  Engine engine = Engine(game: game);
+
+  String formatResult(EngineResult res) {
+    if (!res.hasMove) return 'No Move';
+    String san = game.toSan(res.move!);
+    return '$san (${res.eval}) [depth ${res.depth}]';
+  }
+
+  while (!game.gameOver) {
+    String playerName = game.turn == WHITE ? 'White' : 'Black';
+    print('~~ $playerName is thinking..');
+    EngineResult res = await engine.search();
+    print('Best Move: ${formatResult(res)}');
+    if (res.hasMove) game.makeMove(res.move!);
+    print(game.ascii());
+    print(game.fen);
+  }
+}